requirements.txt
tree-sitter>=0.20.1
torch>=1.9.0
transformers>=4.11.3
faiss-cpu>=1.7.2  # 或 faiss-gpu 用于GPU支持
numpy>=1.19.5
tree-sitter-languages>=1.5.0